requirements.txt
Pillow==6.2.1
pyparsing==2.3.0
torch==1.7.1
torchvision==0.8.2
